# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: events
    description: Public event discovery and scheduling
  - name: registrations
    description: Attendee registrations
  - name: donations
    description: Donation and charity tracking
  - name: analytics
    description: Organizer metrics and activity
  - name: notifications
    description: Event notification preferences
  - name: flyers
    description: Event flyer review and approval
  - name: funds
    description: Community and development funds
  - name: impact
    description: Public giving and development metrics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List public events
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: timeframe
          in: query
          required: false
          schema:
            type: string
            enum: [upcoming, past]
            default: upcoming
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{eventId}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - $ref: "#/components/parameters/EventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /organizer/events:
    get:
      operationId: listOrganizerEvents
      tags: [events]
      summary: List every event for the organizer desk
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [draft, published, completed]
      responses:
        "200":
          description: Events including drafts and completed programs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
  /registrations:
    get:
      operationId: listRegistrations
      tags: [registrations]
      summary: List tickets for an event
      parameters:
        - name: eventId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Registration"
    post:
      operationId: createRegistration
      tags: [registrations]
      summary: Register for an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Registration confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "409":
          description: Registration conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /registrations/{registrationId}:
    get:
      operationId: getRegistration
      tags: [registrations]
      summary: Get a registration
      parameters:
        - $ref: "#/components/parameters/RegistrationId"
      responses:
        "200":
          description: Registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
  /registrations/{registrationId}/check-in:
    post:
      operationId: checkInRegistration
      tags: [registrations]
      summary: Check in a ticket
      parameters:
        - $ref: "#/components/parameters/RegistrationId"
      responses:
        "200":
          description: Checked-in ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
  /flyers:
    get:
      operationId: listFlyers
      tags: [flyers]
      summary: List event flyers
      parameters:
        - name: eventId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Flyers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Flyer"
    post:
      operationId: createFlyer
      tags: [flyers]
      summary: Submit a flyer for approval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlyerInput"
      responses:
        "201":
          description: Flyer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flyer"
  /flyers/{flyerId}:
    patch:
      operationId: updateFlyer
      tags: [flyers]
      summary: Approve or reject a flyer
      parameters:
        - name: flyerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlyerUpdate"
      responses:
        "200":
          description: Flyer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flyer"
  /funds:
    get:
      operationId: listFunds
      tags: [funds]
      summary: List community and development funds
      responses:
        "200":
          description: Funds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fund"
  /funds/{fundId}/gifts:
    post:
      operationId: createFundGift
      tags: [funds]
      summary: Record a gift toward a fund
      parameters:
        - name: fundId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FundGiftInput"
      responses:
        "201":
          description: Updated fund
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fund"
  /impact:
    get:
      operationId: getImpact
      tags: [impact]
      summary: Public giving, tickets, and development metrics
      responses:
        "200":
          description: Public impact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Impact"
  /donations:
    get:
      operationId: listDonations
      tags: [donations]
      summary: List donations
      parameters:
        - name: eventId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Donations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Donation"
    post:
      operationId: createDonation
      tags: [donations]
      summary: Record a donation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonationInput"
      responses:
        "201":
          description: Donation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
  /analytics/overview:
    get:
      operationId: getAnalyticsOverview
      tags: [analytics]
      summary: Get organizer dashboard metrics
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsOverview"
  /analytics/activity:
    get:
      operationId: getAnalyticsActivity
      tags: [analytics]
      summary: Get recent organizer activity
      responses:
        "200":
          description: Activity timeline
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
  /notifications/preferences:
    get:
      operationId: getNotificationPreferences
      tags: [notifications]
      summary: Get notification preferences
      responses:
        "200":
          description: Notification preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
    patch:
      operationId: updateNotificationPreferences
      tags: [notifications]
      summary: Update notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationPreferencesUpdate"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
components:
  parameters:
    EventId:
      name: eventId
      in: path
      required: true
      schema:
        type: integer
    RegistrationId:
      name: registrationId
      in: path
      required: true
      schema:
        type: integer
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Event:
      type: object
      required: [id, title, slug, description, category, startsAt, endsAt, location, capacity, registeredCount, charityName, charityGoal, charityRaised, status]
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        description: { type: string }
        category: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        location: { type: string }
        imageUrl: { type: ["string", "null"] }
        capacity: { type: integer }
        registeredCount: { type: integer }
        charityName: { type: string }
        charityGoal: { type: number }
        charityRaised: { type: number }
        status: { type: string, enum: [draft, published, completed] }
    EventInput:
      type: object
      required: [title, description, category, startsAt, endsAt, location, capacity, charityName, charityGoal]
      properties:
        title: { type: string, minLength: 2 }
        description: { type: string, minLength: 2 }
        category: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        location: { type: string }
        imageUrl: { type: ["string", "null"] }
        capacity: { type: integer, minimum: 1 }
        charityName: { type: string }
        charityGoal: { type: number, minimum: 0 }
    EventUpdate:
      type: object
      properties:
        title: { type: string, minLength: 2 }
        description: { type: string, minLength: 2 }
        category: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        location: { type: string }
        imageUrl: { type: ["string", "null"] }
        capacity: { type: integer, minimum: 1 }
        charityName: { type: string }
        charityGoal: { type: number, minimum: 0 }
        status: { type: string, enum: [draft, published, completed] }
    Registration:
      type: object
      required: [id, eventId, name, email, status, ticketCode, createdAt]
      properties:
        id: { type: integer }
        eventId: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        status: { type: string, enum: [confirmed, waitlisted, cancelled] }
        ticketCode: { type: string }
        checkedInAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    RegistrationInput:
      type: object
      required: [eventId, name, email]
      properties:
        eventId: { type: integer }
        name: { type: string, minLength: 2 }
        email: { type: string, format: email }
    Donation:
      type: object
      required: [id, eventId, donorName, amount, createdAt]
      properties:
        id: { type: integer }
        eventId: { type: integer }
        donorName: { type: string }
        amount: { type: number, minimum: 1 }
        note: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    DonationInput:
      type: object
      required: [eventId, donorName, amount]
      properties:
        eventId: { type: integer }
        donorName: { type: string, minLength: 2 }
        amount: { type: number, minimum: 1 }
        note: { type: ["string", "null"] }
    AnalyticsOverview:
      type: object
      required: [totalEvents, upcomingEvents, totalRegistrations, registrationChange, totalRaised, goalProgress, averageAttendance]
      properties:
        totalEvents: { type: integer }
        upcomingEvents: { type: integer }
        totalRegistrations: { type: integer }
        registrationChange: { type: number }
        totalRaised: { type: number }
        goalProgress: { type: number }
        averageAttendance: { type: number }
    Activity:
      type: object
      required: [id, type, message, createdAt]
      properties:
        id: { type: integer }
        type: { type: string, enum: [registration, donation, event, notification] }
        message: { type: string }
        createdAt: { type: string, format: date-time }
    NotificationPreferences:
      type: object
      required: [emailReminders, registrationAlerts, donationUpdates, weeklyDigest]
      properties:
        emailReminders: { type: boolean }
        registrationAlerts: { type: boolean }
        donationUpdates: { type: boolean }
        weeklyDigest: { type: boolean }
    NotificationPreferencesUpdate:
      type: object
      properties:
        emailReminders: { type: boolean }
        registrationAlerts: { type: boolean }
        donationUpdates: { type: boolean }
        weeklyDigest: { type: boolean }
    Flyer:
      type: object
      required: [id, eventId, title, imageUrl, submittedBy, status, createdAt]
      properties:
        id: { type: integer }
        eventId: { type: integer }
        title: { type: string }
        imageUrl: { type: string }
        submittedBy: { type: string }
        status: { type: string, enum: [pending, approved, rejected] }
        reviewNote: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        reviewedAt: { type: ["string", "null"], format: date-time }
    FlyerInput:
      type: object
      required: [eventId, title, imageUrl, submittedBy]
      properties:
        eventId: { type: integer }
        title: { type: string, minLength: 2 }
        imageUrl: { type: string, minLength: 8 }
        submittedBy: { type: string, minLength: 2 }
    FlyerUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [approved, rejected] }
        reviewNote: { type: ["string", "null"] }
    Fund:
      type: object
      required: [id, name, slug, kind, description, goal, raised]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        kind: { type: string, enum: [program, development] }
        description: { type: string }
        goal: { type: number }
        raised: { type: number }
    FundGiftInput:
      type: object
      required: [donorName, amount]
      properties:
        donorName: { type: string, minLength: 2 }
        amount: { type: number, minimum: 1 }
        note: { type: ["string", "null"] }
    ImpactFund:
      type: object
      required: [id, name, kind, description, goal, raised]
      properties:
        id: { type: integer }
        name: { type: string }
        kind: { type: string, enum: [program, development] }
        description: { type: string }
        goal: { type: number }
        raised: { type: number }
    ImpactGift:
      type: object
      required: [amount, fundName, createdAt]
      properties:
        amount: { type: number }
        fundName: { type: string }
        createdAt: { type: string, format: date-time }
    Impact:
      type: object
      required: [totalRaised, programRaised, developmentRaised, ticketsIssued, ticketsCheckedIn, upcomingEvents, funds, recentGifts]
      properties:
        totalRaised: { type: number }
        programRaised: { type: number }
        developmentRaised: { type: number }
        ticketsIssued: { type: integer }
        ticketsCheckedIn: { type: integer }
        upcomingEvents: { type: integer }
        funds:
          type: array
          items:
            $ref: "#/components/schemas/ImpactFund"
        recentGifts:
          type: array
          items:
            $ref: "#/components/schemas/ImpactGift"